+++ api/src/main/java/io/jsonwebtoken/CompressionCodec.java
@@ -15,6 +15,8 @@
  */
 package io.jsonwebtoken;
 
+import io.jsonwebtoken.io.CompressionAlgorithm;
+
 /**
  * Compresses and decompresses byte arrays according to a compression algorithm.
  *
@@ -27,8 +29,10 @@ package io.jsonwebtoken;
  * @see Jwts.ZIP#DEF
  * @see Jwts.ZIP#GZIP
  * @since 0.6.0
+ * @deprecated since JJWT_RELEASE_VERSION in favor of {@link io.jsonwebtoken.io.CompressionAlgorithm} to equal the RFC name for this concept.
  */
-public interface CompressionCodec extends Identifiable {
+@Deprecated
+public interface CompressionCodec extends CompressionAlgorithm {
 
     /**
      * The algorithm name to use as the JWT
@@ -36,29 +40,10 @@ public interface CompressionCodec extends Identifiable {
      *
      * @return the algorithm name to use as the JWT
      * <a href="https://tools.ietf.org/html/rfc7516#section-4.1.3"><code>zip</code></a> header value.
-     * @deprecated since JJWT_RELEASE_VERSION in favor of {@link Identifiable#getId()} to ensure congruence with
+     * @deprecated since JJWT_RELEASE_VERSION in favor of {@link #getId()} to ensure congruence with
      * all other identifiable algorithms.
      */
     @SuppressWarnings("DeprecatedIsStillUsed")
     @Deprecated
     String getAlgorithmName();
-
-    /**
-     * Compresses the specified byte array, returning the compressed byte array result.
-     *
-     * @param content bytes to compress
-     * @return compressed bytes
-     * @throws CompressionException if the specified byte array cannot be compressed.
-     */
-    byte[] compress(byte[] content) throws CompressionException;
-
-    /**
-     * Decompresses the specified compressed byte array, returning the decompressed byte array result.  The
-     * specified byte array must already be in compressed form.
-     *
-     * @param compressed compressed bytes
-     * @return decompressed bytes
-     * @throws CompressionException if the specified byte array cannot be decompressed.
-     */
-    byte[] decompress(byte[] compressed) throws CompressionException;
 }