+++ api/src/main/java/io/jsonwebtoken/CompressionCodecResolver.java
@@ -15,6 +15,8 @@
  */
 package io.jsonwebtoken;
 
+import java.util.Collection;
+
 /**
  * Looks for a JWT {@code zip} header, and if found, returns the corresponding {@link CompressionCodec} the parser
  * can use to decompress the JWT body.
@@ -23,14 +25,15 @@ package io.jsonwebtoken;
  * {@link Jwts.ZIP#DEF DEFLATE} and {@link Jwts.ZIP#GZIP GZIP} algorithms by default - you do not need to
  * specify a {@code CompressionCodecResolver} in these cases.</p>
  *
- * <p>However, if you want to use a compression algorithm other than {@code DEF} or {@code GZIP}, you must implement
+ * <p>However, if you want to use a compression algorithm other than {@code DEF} or {@code GZIP}, you can implement
  * your own {@link CompressionCodecResolver} and specify that when
- * {@link io.jsonwebtoken.JwtBuilder#compressWith(CompressionCodec) building} and
+ * {@link io.jsonwebtoken.JwtBuilder#compressWith(io.jsonwebtoken.io.CompressionAlgorithm) building} and
  * {@link io.jsonwebtoken.JwtParser#setCompressionCodecResolver(CompressionCodecResolver) parsing} JWTs.</p>
  *
- * @see JwtParserBuilder#setCompressionCodecLocator(Locator)
+ * @see JwtParserBuilder#setCompressionCodecResolver(CompressionCodecResolver)
+ * @see JwtParserBuilder#addCompressionAlgorithms(Collection)
  * @since 0.6.0
- * @deprecated in favor of {@link Locator}
+ * @deprecated in favor of {@link JwtParserBuilder#addCompressionAlgorithms(Collection)}
  */
 @SuppressWarnings("DeprecatedIsStillUsed")
 @Deprecated